Capstone-AI-Powered Customer Support Agent/requirements.txt
langchain
langchain-openai
langgraph
pydantic
python-dotenv
groq
langchain-groq